refactor: 대리점별 카마스터 조회 API 위치 이동
/car-masters API는 카마스터와 대리점 정보를 반환해준다. 해당 API의 위치도 논의해볼 필요 있음
