+++ BE-MyCarMaster/src/main/java/softeer/be_my_car_master/api/car_master/controller/CarMasterController.java
@@ -1,16 +1,18 @@
 package softeer.be_my_car_master.api.car_master.controller;
 
 import javax.validation.Valid;
+import javax.validation.constraints.Min;
 
 import org.springdoc.api.annotations.ParameterObject;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import softeer.be_my_car_master.api.car_master.dto.request.FilterEnum;
+import softeer.be_my_car_master.api.agency.dto.response.GetCarMastersInAgencyResponse;
+import softeer.be_my_car_master.api.car_master.usecase.GetCarMastersInAgencyUseCase;
 import softeer.be_my_car_master.api.car_master.dto.request.GetCarMasterRequest;
 import softeer.be_my_car_master.api.car_master.dto.response.GetCarMasterResponse;
 import softeer.be_my_car_master.api.car_master.usecase.GetCarMasterUseCase;
@@ -18,15 +20,15 @@ import softeer.be_my_car_master.global.exception.BindingParamException;
 import softeer.be_my_car_master.global.response.Response;
 
 @RestController
-@RequestMapping("/car-masters")
 @RequiredArgsConstructor
 @Tag(name = "CarMaster", description = "CarMaster API Document")
 public class CarMasterController {
 
 	private final GetCarMasterUseCase getCarMasterUseCase;
+	private final GetCarMastersInAgencyUseCase getCarMastersInAgencyUseCase;
 
-	@GetMapping
-	public Response<GetCarMasterResponse> getCarMaster(
+	@GetMapping("/car-masters")
+	public Response<GetCarMasterResponse> getCarMasters(
 		@Valid @ParameterObject GetCarMasterRequest getCarMasterRequest,
 		BindingResult bindingResult
 	) {
@@ -39,5 +41,14 @@ public class CarMasterController {
 
 		GetCarMasterResponse getCarMasterResponse = getCarMasterUseCase.execute(latitude, longitude);
 		return Response.createSuccessResponse(getCarMasterResponse);
+	}
+
+	@GetMapping("/agencies/{agencyId}/car-masters")
+	public Response<GetCarMastersInAgencyResponse> getCarMastersInAgency(
+		@Min(value = 1, message = "agencyId는 1 이상의 값입니다.")
+		@PathVariable Long agencyId
+	) {
+		GetCarMastersInAgencyResponse getCarMasterInAgencyResponse = getCarMastersInAgencyUseCase.execute(agencyId);
+		return Response.createSuccessResponse(getCarMasterInAgencyResponse);
 	}
 }
