+++ BE-MyCarMaster/src/test/java/softeer/be_my_car_master/api/car_master/controller/CarMasterControllerTest.java
@@ -8,6 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import java.util.Arrays;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -18,11 +19,14 @@ import org.springframework.test.web.servlet.ResultActions;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import softeer.be_my_car_master.api.agency.dto.response.CarMasterInAgencyDto;
+import softeer.be_my_car_master.api.agency.dto.response.GetCarMastersInAgencyResponse;
 import softeer.be_my_car_master.api.car_master.dto.response.AgencyDto;
 import softeer.be_my_car_master.api.car_master.dto.response.CarMasterAgencyDto;
 import softeer.be_my_car_master.api.car_master.dto.response.CarMasterDto;
 import softeer.be_my_car_master.api.car_master.dto.response.GetCarMasterResponse;
 import softeer.be_my_car_master.api.car_master.usecase.GetCarMasterUseCase;
+import softeer.be_my_car_master.api.car_master.usecase.GetCarMastersInAgencyUseCase;
 import softeer.be_my_car_master.global.response.Response;
 
 @WebMvcTest(CarMasterController.class)
@@ -37,6 +41,14 @@ class CarMasterControllerTest {
 
 	@MockBean
 	private GetCarMasterUseCase getCarMasterUseCase;
+	@MockBean
+	private GetCarMastersInAgencyUseCase getCarMastersInAgencyUseCase;
+
+	@Nested
+	@DisplayName("getCarMaster Test")
+	class GetCarMastersTest {
+
+	}
 
 	@Test
 	@DisplayName("카마스터 목록 조회")
@@ -73,6 +85,38 @@ class CarMasterControllerTest {
 				.contentType(MediaType.APPLICATION_FORM_URLENCODED)
 				.param("latitude", "32.1231")
 				.param("longitude", "127.32333")
+		);
+
+		//then
+		perform
+			.andExpect(status().isOk())
+			.andExpect(content().contentType("application/json"))
+			.andExpect(content().json(responseBody, true));
+	}
+
+	@Test
+	@DisplayName("대리점에 속한 카마스터 목록을 조회합니다")
+	void getCarMastersInAgency() throws Exception {
+		//given
+		GetCarMastersInAgencyResponse getCarMastersInAgencyResponse = new GetCarMastersInAgencyResponse();
+		CarMasterInAgencyDto carMasterInAgencyDto = CarMasterInAgencyDto.builder()
+			.id(1L)
+			.name("신짜오")
+			.imgUrl("imgUrl")
+			.intro("안녕하세요.")
+			.phone("010-0000-0000")
+			.build();
+
+		getCarMastersInAgencyResponse.setCarMasters(Arrays.asList(carMasterInAgencyDto));
+
+		given(getCarMastersInAgencyUseCase.execute(any())).willReturn(getCarMastersInAgencyResponse);
+
+		Response successResponse = Response.createSuccessResponse(getCarMastersInAgencyResponse);
+		String responseBody = objectMapper.writeValueAsString(successResponse);
+
+		//when
+		ResultActions perform = mockMvc.perform(
+			get("/agencies/1/car-masters")
 		);
 
 		//then
